Began integrating the new API
The new graphics API is mostly complete, now I'm actually integrating it into the system. This is a long and slow process.

Most of the file changes appear to be the line endings being modified.

Also added an`ICommandAllocator` type. Turns out this has a lot of benefits. The big one right now is it fixed some potential issues with command queues and modifying a command list after it has been submitted.