Worked on making Reflection Generator more extensible
I haven't integrated it quite yet, but I've been working on making it more of a practical tool as opposed to an MVP or MCVE. I also had to add a new build configuration which allows TauUtils to be built with `/MD` as opposed to `/MT` (dynamic vs static CRT) because my copy of LLVM uses `/MD` and I haven't been able to recompile it. 

I also finally gave in and made PageAllocator initialize itself on its own. I didn't really want to do it, but it shouldn't have any practical impact after a couple of executions thanks to the branch predictor.

The current working idea for the generator is to have attributes for properties. These attributes are implemented through an interface that can parse any amount of tokens in the varargs. This would allow for a wide variety of syntax's such as `foo`, `foo="bar"`, `foo(bar, baz=qux)`. The parser would then emit a pointer to its own internal structure, then this attribute is added to a list of attributes for the property. Currently this "list" is an AVL tree, this will make it easy to search (based on attribute name), and also relatively simple to iterate over. Eventually at a later stage the class that parsed the attribute would be handed the data and expected to emit something.