Mostly finished TauReflectionGenerator
The API is pretty much complete. It's fairly easy to extend with new attributes now. The remaining things to implement are tags for functions, the ability to generate code in the class and not just in the reflection class, and having a proper executable interface that lets you specify the input and output files.