Changed the license to MIT
I've hated licenses that require any "modifications" to use the same license for a while now. Linking against a library is not a modification, and I don't want anyone who thinks otherwise to inhibit the use of this library. Thus I'm switching to the MIT license, one of the most permissive licenses out there.