Worked on resources and string templates
Worked on getting resources working. Still a long way off from completion. In the meantime I started working on adding a string template system for the code gen of TauReflectionGenerator. Thus far I have the grammar, the lexer, some of the AST, and a visitor. Still have to finish the AST and implement the parser.