Added a visitor to repair the String Template AST
Currently the parser doesn't do anything to link together loops and if statements. To make things somewhat simpler I've instead opted to due this later in a visitor. This will go through and connect all of the beginning and endings of loops, link the loop controls to their respective loop, and join together entire if chains.