Worked on shaders and resource mapping.
I've been working on updating shaders to work with the new system, as well as updating how resources are mapped to ensure consistency between mapped data and command list uploads.

Additionally I've redesigned the usage model for resources, moving into a more DX12 style. Now you have 3 types of resource usage heaps: default, upload, and read-back. To get data onto the GPU you have to use an upload resource, to download data you have to use a read-back resource. This is then moved onto or from a default resource by the means of a resource copy in a command list. Internally in DX10 (and most likely DX11 and GL in the future) these upload and read-back resources don't have a backing DX10 type, and instead use simple buffers. Then when uploading it will map a dynamic buffer and copy the data, then unmap the buffer. This allows for maximum efficiency with DX12 resources, while also allowing for a fairly simple pipeline in older API's.